can add meals
Meals can now be added to the meal table. Ingredients can be added to ingredients table. TODO: when ingredients are added for the first time into the table they will not be associated with the meal they were added with. This is an asynchronous issue that needs to be looked into.